requirements.txt
Flask
werkzeug
azure-storage-blob
